feat: export React hooks from library
