+++ framework/hooks/usePopupQuickExit/usePopupQuickExit.mdx
@@ -1,0 +1,65 @@
+---
+name: usePopupQuickExit
+route: /hooks/use-popup-quick-exit
+components: usePopupQuickExit
+---
+
+# usePopupQuickExit
+
+## Install
+
+Integrate with your build to [auto-import](/#integrating), or add an import in your component:
+
+```
+import { usePopupQuickExit } from "@cisco-sbg-ui/magna-react";
+```
+
+## Description
+
+`usePopupQuickExit` is a wrapper around both [`useOutsideClick`](/hooks/use-outside-click) and [`useEscapeKeydown`](/hooks/use-escape-keydown). Often times, when a variation of a popover UI is rendered (such as a modal, drawer, etc.), it should close on _both_ an outside click and an escape keypress. This hook just makes it so that you don't need to import two hooks to handle this behavior.
+
+### Parameters
+
+A configuration object with the following keys:
+
+- `isEnabled`: If the hook should listen for outside clicks and escaspe keypresses
+- `popupRef`: A React Ref to your popup element
+- `onExit`: A function to be called when the user clicks outside of `rootRef` or presses their escape key
+
+## Usage
+
+<Playground
+  code={`() => {
+  const [open, setOpen] = useState(false);
+  const ref = useRef();
+  usePopupQuickExit({
+    popupRef: ref,
+    isEnabled: open,
+    onExit: () => setOpen(false),
+  });
+  return (
+    <div>
+      <AButton onClick={() => setOpen(true)}>Open Modal</AButton>
+      <AModal aria-labelledby="modal-title" isOpen={open}>
+        <APanel ref={ref} style={{ minWidth: "400px" }} type="dialog">
+          <APanelHeader>
+            <APanelTitle id="modal-title">Modal Demo</APanelTitle>
+            <AButton
+              aria-label="Close modal 1"
+              onClick={() => setOpen(false)}
+              tertiaryAlt
+              icon
+            >
+              <AIcon>close</AIcon>
+            </AButton>
+          </APanelHeader>
+          <APanelBody>Modal content goes here.</APanelBody>
+          <APanelFooter>
+            <AButton>Action</AButton>
+          </APanelFooter>
+        </APanel>
+      </AModal>
+    </div>
+  );
+}`}
+/>
